refactor facebook_bot so can test the conversation logic without all the facebook_bot setup
